% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{bdfa}
\alias{bdfa}
\title{Beta DFA}
\usage{
bdfa(data, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Beta DFA estimates
}
\description{
Calculates Beta DFA
}
\keyword{internal}
