% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{betadfa}
\alias{betadfa}
\title{Beta DFA}
\usage{
betadfa(data, dpo = 1, int = TRUE, np = 91, overlap = TRUE)
}
\arguments{
\item{data}{is a matrix of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Beta DFA estimates
}
\description{
Calculates Beta DFA
}
\examples{
set.seed(1)
d <- data.frame(y = cumsum(rnorm(300)), x1 = cumsum(rnorm(300)),
                x2 = cumsum(rnorm(300)))
betadfa(d, np = 20)
}
\references{
Kristoufek, L. (2015). Detrended fluctuation analysis as a regression
framework: estimating dependence at different scales. \emph{Physical Review
E}, 91(2), 022802.
}
\seealso{
\code{\link{sbdfa}}, \code{\link{fracreg}},
\code{vignette("DFATools")}
}
