% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\docType{package}
\name{DFATools-package}
\alias{DFATools}
\alias{DFATools-package}
\title{DFATools: detrended fluctuation analysis and related methods}
\description{
A toolbox of Detrended Fluctuation Analysis ('DFA') methods for long-range
correlations, cross-correlations and regression in nonstationary time series.
See \code{vignette("DFATools")} for the theory and worked examples.
}
\section{Function groups}{

\describe{
  \item{Scaling of a single series}{\code{\link{dfa}}, \code{\link{plotdfa}}}
  \item{Cross-correlation}{\code{\link{rhodcca}}, \code{\link{plotrdcca}},
    \code{\link{plotdcca}}}
  \item{Partial cross-correlation}{\code{\link{rhodpcca}}}
  \item{Multiple cross-correlation}{\code{\link{dmc2}}}
  \item{Detrended fractal regression}{\code{\link{betadfa}},
    \code{\link{sbdfa}}, \code{\link{fracreg}}, \code{\link{effsizeDFA}}}
  \item{Significance tests}{\code{\link{fracreg.PStest}},
    \code{\link{fracreg.Ktest}}, \code{\link{fracreg.IUTest}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Ikarobarreto/DFATools}
  \item Report bugs at \url{https://github.com/Ikarobarreto/DFATools/issues}
}

}
\author{
\strong{Maintainer}: Ikaro Daniel de Carvalho Barreto \email{daniel.carvalho.ib@gmail.com}

Authors:
\itemize{
  \item Luiz Henrique Dore
}

}
\keyword{internal}
