% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{dmc2}
\alias{dmc2}
\title{rho Detrended Multiple-Correlation Coefficient}
\usage{
dmc2(data, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Multiple Detrended Correlation
}
\description{
Calculates DMC²
}
\references{
Zebende, G. F. and da Silva Filho, A. M. (2018). Detrended multiple
cross-correlation coefficient. \emph{Physica A}, 510, 91-97.

Wang, F., Xu, J. and Fan, Q. (2021). Statistical test for detrended multiple
cross-correlation coefficient. \emph{Communications in Nonlinear Science and
Numerical Simulation}, 99, 105781.
}
\seealso{
\code{vignette("DFAmethods2")}
}
