% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{effsizeDFA}
\alias{effsizeDFA}
\title{f² scale-wise effect sizes}
\usage{
effsizeDFA(data, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, f² scale-wise effect sizes
}
\description{
Calculates f² scale-wise effect sizes
}
\references{
Barreto, I. D. C., Dore, L. H., Stosic, T. and Stosic, B. D. (2021).
Extending DFA-based multiple linear regression inference: application to
acoustic impedance models. \emph{Physica A}, 582, 126259.

Cohen, J. (1988). \emph{Statistical Power Analysis for the Behavioral
Sciences}, 2nd ed. Lawrence Erlbaum Associates.
}
\seealso{
\code{\link{fracreg}}, \code{vignette("DFAmethods2")}
}
