% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{fracreg.IUTest}
\alias{fracreg.IUTest}
\title{Intersection-Union Test}
\usage{
fracreg.IUTest(data, B = 100, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series}

\item{B}{number of surrogate series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
A matrix with scale-wise Beta-DFA and critic region of Kristoufek Test and Podobnik-Shen Test.
}
\description{
Calculates Intersection-Union Test for Beta-DFA = 0 or Beta-DFA = Beta-OLS
}
\references{
Barreto, I. D. C., Dore, L. H., Stosic, T. and Stosic, B. D. (2021).
Extending DFA-based multiple linear regression inference: application to
acoustic impedance models. \emph{Physica A}, 582, 126259.
}
\seealso{
\code{\link{fracreg.PStest}}, \code{\link{fracreg.Ktest}},
\code{vignette("DFAmethods2")}
}
