% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{fracreg.Ktest}
\alias{fracreg.Ktest}
\title{Kristoufek Test}
\usage{
fracreg.Ktest(data, B = 100, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series}

\item{B}{number of surrogate series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
A matrix with scale-wise Beta-DFA and critic region of Kristoufek Test.
}
\description{
Calculates Kristoufek Test for Beta-DFA = Beta-OLS
}
