% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{fracreg}
\alias{fracreg}
\title{Multiple Fractal Regression}
\usage{
fracreg(data, dpo, int, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series.}

\item{dpo}{Detrending polynomial order.}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Detrended Fluctuation Function F, Rho-DCCA, Rho DPCCA,
Beta DFA estimates, Standardized Beta DFA estimates, DFA Residuals, DFA Variance, DFA Upper and Lower confidence interval,
Multiple Detrended Correlation, DFA R² , DFA p-value and DFA Calculated T statistics.
}
\description{
Calculates Fractal Regression.
}
