% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{plotdcca}
\alias{plotdcca}
\title{Plot DCCA}
\usage{
plotdcca(dcca, seg = FALSE, point = NULL, main = NULL)
}
\arguments{
\item{dcca}{is a fracreg object}

\item{seg}{logical. If TRUE, alpha DCCA will be calculated in 2 segments}

\item{point}{indicate in which point segmented alpha DFA should be calculated}

\item{main}{plot title}
}
\value{
a plot of Detrended Cross Correlation Analysis.
}
\description{
Plot of Detrended Cross Correlation Analysis
}
\examples{
# 'dcca' is a list with the box scales and the detrended covariance F^2_XY(s)
dcca <- list(s = c(10, 20, 40, 80, 160), Fxy = c(50, 130, 320, 780, 1900))
plotdcca(dcca)
}
