% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{plotdfa}
\alias{plotdfa}
\title{Plot DFA}
\usage{
plotdfa(dfa, seg = F, point = NULL, main)
}
\arguments{
\item{dfa}{is a dfa object}

\item{seg}{logical. If TRUE, alpha DFA will be calculated in 2 segments}

\item{point}{indicate in which point segmented alpha DFA should be calculated}

\item{main}{plot title}
}
\value{
a plot of Detrended Fluctuation Analysis.
}
\description{
Plot of Detrended Fluctuation Analysis
}
