% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{rhodcca}
\alias{rhodcca}
\title{rho Detrended Cross-Correlation Coefficient}
\usage{
rhodcca(data, dpo = 1, int = TRUE, np = 91, overlap = TRUE)
}
\arguments{
\item{data}{is a matrix of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Rho-DCCA
}
\description{
Calculates rho DCCA
}
\examples{
set.seed(1)
d <- data.frame(x = cumsum(rnorm(300)), y = cumsum(rnorm(300)))
rhodcca(d, np = 20)
}
\references{
Podobnik, B. and Stanley, H. E. (2008). Detrended cross-correlation analysis:
a new method for analyzing two nonstationary time series.
\emph{Physical Review Letters}, 100(8), 084102.

Zebende, G. F. (2011). DCCA cross-correlation coefficient: quantifying level
of cross-correlation. \emph{Physica A}, 390(4), 614-618.

Kwapien, J., Oswiecimka, P. and Drozdz, S. (2015). Detrended fluctuation
analysis made flexible to detect range of cross-correlated fluctuations.
\emph{Physical Review E}, 92(5), 052815.

Cavalcanti, S. L. (2019). \emph{Aplicacoes de DFA, DCCA e DPCCA em focos
de calor na Amazonia Legal} (PhD thesis). Universidade Federal Rural de
Pernambuco.
}
\seealso{
\code{vignette("DFATools")}
}
