% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{rhodpcca}
\alias{rhodpcca}
\title{rho Detrended Partial Cross-Correlation Coefficient}
\usage{
rhodpcca(data, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Rho DPCCA
}
\description{
Calculates DPCCA
}
\references{
Yuan, N. et al. (2015). Detrended partial-cross-correlation analysis: a new
method for analyzing correlations in complex system. \emph{Scientific
Reports}, 5, 8143.
}
\seealso{
\code{\link{rhodcca}}, \code{vignette("DFAmethods2")}
}
