% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{dfa}
\alias{dfa}
\title{Detrended Fluctuation Analysis}
\usage{
dfa(data, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a vector of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Detrended Fluctuation Function F
}
\description{
Calculates DFA
}
\references{
Peng, C.-K. et al. (1994). Mosaic organization of DNA nucleotides.
\emph{Physical Review E}, 49(2), 1685-1689.
}
\seealso{
\code{vignette("DFAmethods2")}
}
