% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{fracreg}
\alias{fracreg}
\title{Multiple Fractal Regression}
\usage{
fracreg(data, dpo, int, np = 91, overlap = TRUE)
}
\arguments{
\item{data}{is a matrix of time series.}

\item{dpo}{Detrending polynomial order.}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Detrended Fluctuation Function F, Rho-DCCA, Rho DPCCA,
Beta DFA estimates, Standardized Beta DFA estimates, DFA Residuals, DFA Variance, DFA Upper and Lower confidence interval,
Multiple Detrended Correlation, DFA R² , DFA p-value and DFA Calculated T statistics.
}
\description{
Calculates Fractal Regression.
}
\examples{
set.seed(1)
d <- data.frame(y = cumsum(rnorm(300)), x1 = cumsum(rnorm(300)),
                x2 = cumsum(rnorm(300)))
fracreg(d, dpo = 1, int = TRUE, np = 20)
}
\references{
Barreto, I. D. C., Dore, L. H., Stosic, T. and Stosic, B. D. (2021).
Extending DFA-based multiple linear regression inference: application to
acoustic impedance models. \emph{Physica A}, 582, 126259.

Shen, C. (2015). A new detrended semipartial cross-correlation analysis.
\emph{Physics Letters A}, 379(44), 2962-2969.

Kristoufek, L. (2015). Detrended fluctuation analysis as a regression
framework: estimating dependence at different scales. \emph{Physical Review
E}, 91(2), 022802.
}
\seealso{
\code{\link{betadfa}}, \code{\link{effsizeDFA}},
\code{vignette("DFAmethods2")}
}
