% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{plotrdcca}
\alias{plotrdcca}
\title{Plot rho-DCCA}
\usage{
plotrdcca(rdcca, var)
}
\arguments{
\item{rdcca}{is a rhodcca object}

\item{var}{character. Indicate which pair in rho dcca object you want to plot.}
}
\value{
a plot of Detrended Cross Correlation Analysis.
}
\description{
Plot of Detrended Cross Correlation Coefficient
}
