% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{sbdfa}
\alias{sbdfa}
\title{Standardized Beta DFA}
\usage{
sbdfa(data, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Standardized Beta DFA estimates
}
\description{
Calculates Standardized Beta DFA
}
