% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFAmethods.R
\name{sbdfa}
\alias{sbdfa}
\title{Standardized Beta DFA}
\usage{
sbdfa(data, dpo = 1, int = T, np = 91, overlap = T)
}
\arguments{
\item{data}{is a matrix of time series}

\item{dpo}{detrending polynomial order}

\item{int}{logical. if TRUE integration process will be applied.}

\item{np}{number of point scales.}

\item{overlap}{logical. if TRUE overlapping windows will be applied.}
}
\value{
Scale s, Standardized Beta DFA estimates
}
\description{
Calculates Standardized Beta DFA
}
\references{
Barreto, I. D. C., Dore, L. H., Stosic, T. and Stosic, B. D. (2021).
Extending DFA-based multiple linear regression inference: application to
acoustic impedance models. \emph{Physica A}, 582, 126259.

Kristoufek, L. (2015). Detrended fluctuation analysis as a regression
framework: estimating dependence at different scales. \emph{Physical Review
E}, 91(2), 022802.
}
\seealso{
\code{\link{betadfa}}, \code{vignette("DFAmethods2")}
}
